+++ src/scripts/events.ts
@@ -1,19 +1,27 @@
+import type { Advanced } from '../types/sync.ts'
+
 import { isTypingTarget } from './features/links/helpers.ts'
 
 let isMousingDownOnInput = false
 
-export function userActions(): void {
+export function userActions(advanced: Advanced): void {
     document.body.addEventListener('mousedown', detectTargetAsInputs)
     document.getElementById('b_editmove')?.addEventListener('click', closeSettingsOnMoveOpen)
 
     document.addEventListener('click', clickUserActions)
-    document.addEventListener('keydown', keyboardUserActions)
-    document.addEventListener('keyup', keyboardUserActions)
+
+    document.addEventListener('keydown', (event) => {
+        keyboardUserActions(advanced, event)
+    })
+
+    document.addEventListener('keyup', (event) => {
+        keyboardUserActions(advanced, event)
+    })
 }
 
 // Main functions
 
-function keyboardUserActions(event: KeyboardEvent): void {
+function keyboardUserActions(advanced: Advanced, event: KeyboardEvent): void {
     const { altKey, ctrlKey, metaKey, code, type } = event
 
     const domsuggestions = document.getElementById('sb-suggestions')
@@ -30,7 +38,7 @@ function keyboardUserActions(event: KeyboardEvent): void {
         if (open.contextmenu) {
             document.dispatchEvent(new Event('close-edit'))
         } //
-        else if (open.settings && keyup) {
+        else if (advanced.escKey && open.settings && keyup) {
             document.dispatchEvent(new CustomEvent('toggle-settings'))
         } //
         else if (open.selectall) {
@@ -39,7 +47,7 @@ function keyboardUserActions(event: KeyboardEvent): void {
         else if (open.folder) {
             document.dispatchEvent(new Event('close-folder'))
         } //
-        else if (keyup) {
+        else if (advanced.escKey && keyup) {
             // condition to avoid conflicts with esc key on supporters modal
             // likely to be improved
             if (document.documentElement.dataset.supportersModal === undefined) {